sql: move SHOW handlers to new module
The plan is to eventually split up the monolithic statement module
completely.
